# docs/openapi-v1.yml
openapi: 3.0.3
info:
  title: Selene API — v1
  description: >
    Public REST contract for the Selene smart energy & climate platform.
    Served by the Caddy gateway under the `/api/v1` prefix; each path below is
    routed to a dedicated microservice (services/auth, services/energy,
    services/climate, services/analytics, services/firmware).

    Versioning lives at the gateway: `handle_path /api/v1/*` strips the prefix
    before proxying, so services define version-agnostic paths (`/energy/latest`)
    and a future `/api/v2` needs no service changes.

    Auth: short-lived EdDSA (Ed25519) access tokens with rotating refresh
    tokens (one-time use, reuse detection). JWKS at `/.well-known/jwks.json`.
  version: 1.0.0
  contact:
    name: Selene
  license:
    name: Proprietary
    url: https://github.com/dankehidayat/selene
servers:
  - url: https://selene.dankehidayat.my.id/api/v1
    description: Production
  - url: http://localhost:3009/api/v1
    description: Local (gateway / auth)
...
paths:
...
  /auth/2fa/status:
    get:

      operationId: getTwoFactorStatus
      tags: [auth]
      summary: Whether the current user has 2FA enabled
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: 2FA enablement status.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  enabled:
                    type: boolean
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/2fa/setup:
    post:

      operationId: setupTwoFactor
      tags: [auth]
      summary: Generate a TOTP setup (returns provisioning secret + otpauth URI)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Pending setup token carries the secret for the verify step.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFactorSetup"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/2fa/enable:
    post:

      operationId: enableTwoFactor
      tags: [auth]
      summary: Confirm a TOTP secret by code and enable 2FA
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorSetupVerifyRequest"
      responses:
        "200":
          description: 2FA enabled with fresh backup codes.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  enabled:
                    type: boolean
                  backupCodes:
                    type: array
                    items:
                      type: string
        "400":
          description: Wrong code or missing setup token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /auth/2fa/disable:
    post:

      operationId: disableTwoFactor
      tags: [auth]
      summary: Disable 2FA (requires password + current TOTP/backup code)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorVerifyRequest"
      responses:
        "200":
          description: 2FA disabled.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  enabled:
                    type: boolean
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/2fa/backup-codes:
    post:

      operationId: regenerateBackupCodes
      tags: [auth]
      summary: Regenerate backup codes (requires password + current TOTP code)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFactorVerifyRequest"
      responses:
        "200":
          description: Fresh backup codes (show once).
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  backupCodes:
                    type: array
                    items:
                      type: string
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived EdDSA (Ed25519) access token; refresh via POST /auth/refresh. Verify against /.well-known/jwks.json.
...
  responses:
    BadRequest:
      description: Invalid payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing, invalid, or expired token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, RATE_LIMITED, VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR, TOKEN_REUSE_DETECTED]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
...
    TwoFactorVerifyRequest:
      type: object
      additionalProperties: false
      required: [code, password]
      properties:
        code:
          type: string
          description: TOTP or backup code for confirmation/disable.
        password:
          type: string
          format: password
          description: Account password (reconfirmation before disable/regeneration).
    TwoFactorSetupVerifyRequest:
      type: object
      additionalProperties: false
      required: [pendingToken, code]
      properties:
        pendingToken:
          type: string
          description: Token returned by 2FA setup.
        code:
          type: string
          description: Current TOTP code proving access to the secret.
    TwoFactorSetup:
      type: object
      additionalProperties: false
      properties:
        pendingToken:
          type: string
        secret:
          type: string
        otpauthUri:
          type: string
...